+++ ChangeLog.md
@@ -1,5 +1,10 @@
 # What's new in the WSL Management PowerShell module
 
+## Version 2.1 (2023-09-26)
+
+- Added support for retrieving WSL distribution listing from online sources via
+  [`Get-WslDistributionOnline`][] cmdlet.
+
 ## Version 2.0 (2023-09-11)
 
 - Improvements to the [`Import-WslDistribution`][] and [`Export-WslDistribution`][] cmdlets.
@@ -35,3 +40,4 @@
 [`Get-WslVersion`]: docs/Get-WslVersion.md
 [`Import-WslDistribution`]: docs/Import-WslDistribution.md
 [`Invoke-WslCommand`]: docs/Invoke-WslCommand.md
+[`Get-WslDistributionOnline`]: docs/Get-WslDistributionOnline.md